What: refactor RunMode type
Why:

  * migration to const.go

How:

  * nop
